Authorization With Policies

+++ app/Http/Controllers/ThreadsController.php
@@ -102,6 +102,9 @@ class ThreadsController extends Controller
      */
     public function destroy(Channel $channel, Thread $thread)
     {
+        $this->authorize('delete',$thread);
+
+
         $thread->delete();
 
         if(request()->wantsJson())
